# lib/api-spec/openapi.yml
openapi: 3.1.0
info:
  title: TestCaseHub API
  version: "1.0.0"
  description: UAT Test Case Management API
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: integer }
        project_code: { type: string }
        name: { type: string }
        designed_by: { type: string }
        module_name: { type: string }
        design_date: { type: string }
        test_link: { type: string, nullable: true }
        version: { type: integer }
        version_date: { type: string, format: date-time }
        test_lead_id: { type: integer, nullable: true }
        is_signed_off: { type: integer }
        sign_off_data: { type: string, nullable: true }
        objectives: { type: string, nullable: true }
        scope: { type: string, nullable: true }
        out_of_scope: { type: string, nullable: true }
        entry_criteria: { type: string, nullable: true }
        exit_criteria: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    TestCase:
      type: object
      properties:
        id: { type: integer }
        use_case_id: { type: integer }
        case_number: { type: string }
        title: { type: string }
        test_type: { type: string, nullable: true }
        estimated_minutes: { type: integer, nullable: true }
        acceptance_criteria: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      summary: Create project (Admin only)
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                designedBy: { type: string }
                moduleName: { type: string }
                designDate: { type: string }
                testLink: { type: string, nullable: true }
                testLeadId: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      summary: Get project by id
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    put:
      summary: Update project
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "200":
          description: OK
    delete:
      summary: Delete project (Admin only)
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /users/{userId}/projects:
    get:
      summary: List user's projects
      tags: [ProjectAssignments]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
...
  /use-cases/{useCaseId}/test-cases:
    get:
      summary: List test cases for use case
      tags: [TestCases]
      parameters:
        - in: path
          name: useCaseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      summary: Create test case
      tags: [TestCases]
      parameters:
        - in: path
          name: useCaseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caseNumber: { type: string }
                title: { type: string }
                testType: { type: string, nullable: true }
                estimatedMinutes: { type: integer, nullable: true }
                acceptanceCriteria: { type: string, nullable: true }
      responses:
        "201":
          description: Created
  /test-cases/{testCaseId}:
    put:
      summary: Update test case
      tags: [TestCases]
      parameters:
        - in: path
          name: testCaseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCase"
      responses:
        "200":
          description: OK
    delete:
      summary: Delete test case
      tags: [TestCases]
      parameters:
        - in: path
          name: testCaseId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /projects/{projectId}/sign-off:
    post:
      summary: Sign off project
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /projects/{projectId}/sign-off-status:
    get:
      summary: Get sign off status
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /projects/{projectId}/uat-summary:
    get:
      summary: UAT summary
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /projects/{projectId}/audit-log:
    get:
      summary: Audit log for project
      tags: [Projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
  /projects/code/{projectCode}:
    get:
      summary: Get project by code
      tags: [Projects]
      parameters:
        - in: path
          name: projectCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
security:
  - BearerAuth: []